# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: UPCAST Data Profiling API
  description: OpenAPI specification for the UPCAST Data Profiling API
  version: 0.1.0
paths:
  /profile/get_profilers:
    get:
      summary: Get Available Profilers
      operationId: get_available_profilers_profile_get_profilers_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  /profile/generate_profile:
    post:
      summary: Generate Profile For Local File With Selected Profiler
      operationId: generate_profile_for_local_file_with_selected_profiler_profile_generate_profile_post
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_generate_profile_for_local_file_with_selected_profiler_profile_generate_profile_post'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /profile/get_profile:
    get:
      summary: Get Profile From Url
      operationId: get_profile_from_url_profile_get_profile_get
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            title: Url
        - name: profiler_name
          in: query
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Profiler Name
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /profile/update_profile:
    put:
      summary: Update Profile Specified By Url With Local File
      operationId: update_profile_specified_by_url_with_local_file_profile_update_profile_put
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            title: Url
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_update_profile_specified_by_url_with_local_file_profile_update_profile_put'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /profile/delete_profile:
    delete:
      summary: Delete Profile Specified By Url
      operationId: delete_profile_specified_by_url_profile_delete_profile_delete
      parameters:
        - name: url
          in: query
          required: true
          schema:
            type: string
            title: Url
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    Body_generate_profile_for_local_file_with_selected_profiler_profile_generate_profile_post:
      properties:
        selected_app:
          type: string
          title: Selected App
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
        - selected_app
        - file
      title: Body_generate_profile_for_local_file_with_selected_profiler_profile_generate_profile_post
    Body_update_profile_specified_by_url_with_local_file_profile_update_profile_put:
      properties:
        file_content:
          type: string
          format: binary
          title: File Content
          description: Local file to update with
      type: object
      required:
        - file_content
      title: Body_update_profile_specified_by_url_with_local_file_profile_update_profile_put
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError